program-runtime: port over some tests from syscalls lib
